+++ lib/ethon/easy.rb
@@ -179,6 +179,10 @@ module Ethon
       Curl.easy_reset(handle)
     end
 
+    def escape(value)
+      Curl.easy_escape(handle, value, 0)
+    end
+
     # Returns a  pointer to the curl easy handle.
     #
     # @example Return the handle.